Add files via upload
This is a Jupyter Notebook file with extension .ipynb and it includes Data preprocessing, Neural Network, LSTM and CNN for Stock prediction of dataset provided from some random company. 
This project is threefold:

Task 1: Use the daily [Open, High, Low, Volume] to predict [Close] on that day using a fully-connected neural network. Use the first 70% of the records for training and the remaining 30% of the records for test. Report the RMSE of the model. Show the “regression lift chart” of your test data.

Task 2: Predict [Close] of a day based on the last 7 days’ data [Open, High, Low, Volume, Close] using a LSTM model. In other words, we want to predict the price in the green cell using all the numbers in the red cell. Use the first 70% of the available records for training and the remaining 30% of the available records for test. Report the RMSE of the model. Show the “regression lift chart” of your test data.

Task 3: Do the same as Task 2 but use a CNN model. Report the RMSE of the model. Show the “regression lift chart” of your test data.